Show updating indicator on rooms list
